CORE: Move emitters to a dynamic plugin system
I've been meaning to do this for literally years.  This just cleans up
the interface for adding new emitters.  Now instead of having to
touch a bunch of files all over the place, you just need to dump your
emitter file and any related templates in the proper directories and
you're off to the races