CORE|C: Allow partial encode/decode functionality
This has been on the roadmap for a while, and a half-ass
implementatino existed in the original several years ago, but this new
version is much closer to what I originally envisioned.

Now, instead of _target fields in each message, there are two fields:
_encoders and _decoders.

Depending on the emitter, these fields may be optionally used to only
generate code for a subset of messages. For example, in a request
response scenario, request messages may only be decoded by the
requestee, and responses may only be encoded (not decoded).  While
this may seem odd, on resource constrained microcontrollers, the
ability to selectively omit code generation is important for flash
space constraints.

Note that these fields are optional and some targets may simply opt to
generate everything (python and Nim for example) as they are not
resource constrained.